live reloading of calories for potatos.
